pyproject.toml
[tool.poetry]
name = "mtl-langchain"
version = "0.1.0"
description = ""
authors = ["Mariusz Smenżyk <mariusz@bravelab.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.7"
langchain-openai = "^0.2.8"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
